precis_core

Module profile

This module contains traits with operations and rules that profiles must implement such as it is defined by the PRECIS framework rfc8264

Traits

PrecisFastInvocation

Fast invocation trait that allows profiles to be used without providing a specific instance. This is usually achieved by using a static instance allocated with lazy_static

Profile

Profile enables application protocols to apply the string classes in ways that are appropriate for common constructs.

Rules

Rules that any profile of a PRECIS string class MUST define to proper manage the handling of right-to-left code points as well as various mapping operations such as case preservation or lower casing, Unicode normalization, mapping of certain code points to other code points or to nothing, and mapping of full width and half width code points.

Functions

stabilize

Apply rules until the string is stable. Some profiles, especially those that the result of applying these rules does not result in an idempotent operation for all code points SHOULD apply the rules repeatedly until the output string is stable.