precis_core::profile

Trait Rules

pub trait Rules {
    // Provided methods
    fn width_mapping_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
       where T: Into<Cow<'a, str>> { ... }
    fn additional_mapping_rule<'a, T>(
        &self,
        _s: T,
    ) -> Result<Cow<'a, str>, Error>
       where T: Into<Cow<'a, str>> { ... }
    fn case_mapping_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
       where T: Into<Cow<'a, str>> { ... }
    fn normalization_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
       where T: Into<Cow<'a, str>> { ... }
    fn directionality_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
       where T: Into<Cow<'a, str>> { ... }
}

Rules that any profile of a PRECIS string class MUST define to proper manage the handling of right-to-left code points as well as various mapping operations such as case preservation or lower casing, Unicode normalization, mapping of certain code points to other code points or to nothing, and mapping of full width and half width code points.

Provided Methods

fn width_mapping_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
where T: Into<Cow<'a, str>>,

Applies the width mapping rule of a profile to an input string.

Arguments:

s: String value

Returns

The same string if no modifications were required or a new allocated string if s was modified as a result of applying this rule

fn additional_mapping_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
where T: Into<Cow<'a, str>>,

Applies the additional mapping rule of a profile to an input string.

Arguments:

s: String value

Returns

The same string if no modifications were required or a new allocated string if s was modified as a result of applying this rule

fn case_mapping_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
where T: Into<Cow<'a, str>>,

Applies the case mapping rule of a profile to an input string

Arguments:

s: String value

Returns

The same string if no modifications were required or a new allocated string if s was modified as a result of applying this rule

fn normalization_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
where T: Into<Cow<'a, str>>,

Applies the normalization rule of a profile to an input string

Arguments:

s: String value

Returns

The same string if no modifications were required or a new allocated string if s was modified as a result of applying this rule

fn directionality_rule<'a, T>(&self, _s: T) -> Result<Cow<'a, str>, Error>
where T: Into<Cow<'a, str>>,

Applies the directionality rule of a profile to an input string

Arguments:

s: String value

Returns

The same string if no modifications were required or a new allocated string if s was modified as a result of applying this rule

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.