precis_core::stringclasses

Trait SpecificDerivedPropertyValue

pub trait SpecificDerivedPropertyValue {
    // Required methods
    fn on_spaces(&self) -> DerivedPropertyValue;
    fn on_symbols(&self) -> DerivedPropertyValue;
    fn on_punctuation(&self) -> DerivedPropertyValue;
    fn on_has_compat(&self) -> DerivedPropertyValue;
    fn on_other_letter_digits(&self) -> DerivedPropertyValue;
}

Interface for specific classes to deal with specific Unicode code groups defined in RFC 8264. Next callbacks will be invoked to calculate the derived property according to the algorithm defined in rfc8264

Required Methods

fn on_spaces(&self) -> DerivedPropertyValue

Callback invoked when the Unicode code point belongs to Spaces

fn on_symbols(&self) -> DerivedPropertyValue

Callback invoked when the Unicode code point belongs to Symbols

fn on_punctuation(&self) -> DerivedPropertyValue

Callback invoked when the Unicode code point belongs to Punctuation

fn on_has_compat(&self) -> DerivedPropertyValue

Callback invoked when the Unicode code point belongs to HasCompat

fn on_other_letter_digits(&self) -> DerivedPropertyValue

Callback invoked when the Unicode code point belongs to OtherLetterDigits